# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SkillLevelInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  skill: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  current: number;
}

export interface SkillTwinAnalysisInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  targetRole: string;
  /** @maxLength 16000 */
  resumeText?: string;
  /**
     * @minItems 5
     * @maxItems 7
     */
  skillLevels?: SkillLevelInput[];
}

export interface SkillGap {
  /** @maxLength 80 */
  skill: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  current: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  required: number;
  /** @maxLength 320 */
  rationale: string;
  /**
     * @maxItems 6
     * @items.maxLength 80
     */
  missingTools?: string[];
}

export interface WeeklyMilestone {
  /** @maxLength 30 */
  week: string;
  /** @maxLength 320 */
  goal: string;
  /** @maxLength 220 */
  deliverable: string;
}

export interface LearningResource {
  /** @maxLength 140 */
  title: string;
  /** @maxLength 100 */
  provider: string;
  /** @maxLength 220 */
  reason: string;
  free: boolean;
}

export type RoadmapPhasePhase = typeof RoadmapPhasePhase[keyof typeof RoadmapPhasePhase];


export const RoadmapPhasePhase = {
  '30_days': '30 days',
  '60_days': '60 days',
  '90_days': '90 days',
} as const;

export interface RoadmapPhase {
  phase: RoadmapPhasePhase;
  /** @maxLength 100 */
  title: string;
  /** @maxLength 220 */
  focus: string;
  /**
     * @minItems 4
     * @maxItems 4
     */
  weeklyMilestones: WeeklyMilestone[];
  /** @maxLength 360 */
  portfolioProject: string;
  /**
     * @minItems 2
     * @maxItems 3
     */
  resources: LearningResource[];
  /**
     * @minItems 1
     * @maxItems 3
     * @items.maxLength 260
     */
  resumeTips: string[];
}

export type SkillTwinAnalysisSource = typeof SkillTwinAnalysisSource[keyof typeof SkillTwinAnalysisSource];


export const SkillTwinAnalysisSource = {
  ai: 'ai',
  local: 'local',
} as const;

export interface SkillTwinAnalysis {
  targetRole: string;
  source: SkillTwinAnalysisSource;
  /**
     * @minimum 0
     * @maximum 100
     */
  readiness: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  matchScore: number;
  /** @maxLength 800 */
  summary: string;
  /**
     * @minItems 1
     * @maxItems 6
     * @items.maxLength 320
     */
  strengths: string[];
  /**
     * @minItems 1
     * @maxItems 6
     * @items.maxLength 320
     */
  shortfalls: string[];
  /**
     * @minItems 5
     * @maxItems 7
     */
  gaps: SkillGap[];
  /**
     * @minItems 3
     * @maxItems 3
     */
  roadmap: RoadmapPhase[];
}

export interface MockInterviewQuestionsInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  targetRole: string;
  /** @maxLength 100 */
  targetCompany?: string;
}

export type MockInterviewQuestionsSource = typeof MockInterviewQuestionsSource[keyof typeof MockInterviewQuestionsSource];


export const MockInterviewQuestionsSource = {
  ai: 'ai',
  local: 'local',
} as const;

export interface MockInterviewQuestions {
  source: MockInterviewQuestionsSource;
  targetRole: string;
  targetCompany: string;
  /**
     * @minItems 3
     * @maxItems 5
     * @items.maxLength 500
     */
  questions: string[];
}

export interface InterviewAnswer {
  /** @maxLength 500 */
  question: string;
  /** @maxLength 5000 */
  answer: string;
}

export interface MockInterviewEvaluationInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  targetRole: string;
  /** @maxLength 100 */
  targetCompany?: string;
  /**
     * @minItems 3
     * @maxItems 5
     */
  responses: InterviewAnswer[];
}

export type MockInterviewEvaluationSource = typeof MockInterviewEvaluationSource[keyof typeof MockInterviewEvaluationSource];


export const MockInterviewEvaluationSource = {
  ai: 'ai',
  local: 'local',
} as const;

export interface MockInterviewEvaluation {
  source: MockInterviewEvaluationSource;
  targetRole: string;
  targetCompany: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  overallScore: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  clarityScore: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  technicalAccuracy: number;
  /**
     * @maxItems 10
     * @items.maxLength 100
     */
  conceptsCovered: string[];
  /**
     * @maxItems 10
     * @items.maxLength 100
     */
  missedConcepts: string[];
  /** @maxLength 700 */
  summary: string;
  /**
     * @minItems 2
     * @maxItems 8
     * @items.maxLength 320
     */
  tips: string[];
}

export interface ApiError {
  error: string;
}