# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: skilltwin
    description: AI-powered student career readiness analysis
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /skilltwin/analyze:
    post:
      operationId: analyzeSkillTwin
      tags: [skilltwin]
      summary: Analyze readiness for any target role
      description: Generates role-specific skill requirements, detailed career feedback, and a 30/60/90-day plan.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillTwinAnalysisInput"
      responses:
        "200":
          description: Role-specific readiness analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillTwinAnalysis"
        "400":
          description: Invalid analysis request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /skilltwin/interview/questions:
    post:
      operationId: generateMockInterview
      tags: [skilltwin]
      summary: Generate a role and company specific mock interview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MockInterviewQuestionsInput"
      responses:
        "200":
          description: Interview questions ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MockInterviewQuestions"
        "400":
          description: Invalid interview request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /skilltwin/interview/evaluate:
    post:
      operationId: evaluateMockInterview
      tags: [skilltwin]
      summary: Evaluate mock interview answers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MockInterviewEvaluationInput"
      responses:
        "200":
          description: Interview feedback ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MockInterviewEvaluation"
        "400":
          description: Invalid interview evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SkillLevelInput:
      type: object
      properties:
        skill:
          type: string
          minLength: 1
          maxLength: 80
        current:
          type: integer
          minimum: 0
          maximum: 100
      required: [skill, current]
    SkillTwinAnalysisInput:
      type: object
      properties:
        targetRole:
          type: string
          minLength: 2
          maxLength: 100
        resumeText:
          type: string
          maxLength: 16000
        skillLevels:
          type: array
          items:
            $ref: "#/components/schemas/SkillLevelInput"
          minItems: 5
          maxItems: 7
      required: [targetRole]
    SkillGap:
      type: object
      properties:
        skill:
          type: string
          maxLength: 80
        current:
          type: integer
          minimum: 0
          maximum: 100
        required:
          type: integer
          minimum: 0
          maximum: 100
        rationale:
          type: string
          maxLength: 320
        missingTools:
          type: array
          items:
            type: string
            maxLength: 80
          maxItems: 6
      required: [skill, current, required, rationale]
    WeeklyMilestone:
      type: object
      properties:
        week:
          type: string
          maxLength: 30
        goal:
          type: string
          maxLength: 320
        deliverable:
          type: string
          maxLength: 220
      required: [week, goal, deliverable]
    LearningResource:
      type: object
      properties:
        title:
          type: string
          maxLength: 140
        provider:
          type: string
          maxLength: 100
        reason:
          type: string
          maxLength: 220
        free:
          type: boolean
      required: [title, provider, reason, free]
    RoadmapPhase:
      type: object
      properties:
        phase:
          type: string
          enum: ["30 days", "60 days", "90 days"]
        title:
          type: string
          maxLength: 100
        focus:
          type: string
          maxLength: 220
        weeklyMilestones:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyMilestone"
          minItems: 4
          maxItems: 4
        portfolioProject:
          type: string
          maxLength: 360
        resources:
          type: array
          items:
            $ref: "#/components/schemas/LearningResource"
          minItems: 2
          maxItems: 3
        resumeTips:
          type: array
          items:
            type: string
            maxLength: 260
          minItems: 1
          maxItems: 3
      required: [phase, title, focus, weeklyMilestones, portfolioProject, resources, resumeTips]
    SkillTwinAnalysis:
      type: object
      properties:
        targetRole:
          type: string
        source:
          type: string
          enum: [ai, local]
        readiness:
          type: integer
          minimum: 0
          maximum: 100
        matchScore:
          type: integer
          minimum: 0
          maximum: 100
        summary:
          type: string
          maxLength: 800
        strengths:
          type: array
          items:
            type: string
            maxLength: 320
          minItems: 1
          maxItems: 6
        shortfalls:
          type: array
          items:
            type: string
            maxLength: 320
          minItems: 1
          maxItems: 6
        gaps:
          type: array
          items:
            $ref: "#/components/schemas/SkillGap"
          minItems: 5
          maxItems: 7
        roadmap:
          type: array
          items:
            $ref: "#/components/schemas/RoadmapPhase"
          minItems: 3
          maxItems: 3
      required: [targetRole, source, readiness, matchScore, summary, strengths, shortfalls, gaps, roadmap]
    MockInterviewQuestionsInput:
      type: object
      properties:
        targetRole:
          type: string
          minLength: 2
          maxLength: 100
        targetCompany:
          type: string
          maxLength: 100
      required: [targetRole]
    MockInterviewQuestions:
      type: object
      properties:
        source:
          type: string
          enum: [ai, local]
        targetRole:
          type: string
        targetCompany:
          type: string
        questions:
          type: array
          items:
            type: string
            maxLength: 500
          minItems: 3
          maxItems: 5
      required: [source, targetRole, targetCompany, questions]
    InterviewAnswer:
      type: object
      properties:
        question:
          type: string
          maxLength: 500
        answer:
          type: string
          maxLength: 5000
      required: [question, answer]
    MockInterviewEvaluationInput:
      type: object
      properties:
        targetRole:
          type: string
          minLength: 2
          maxLength: 100
        targetCompany:
          type: string
          maxLength: 100
        responses:
          type: array
          items:
            $ref: "#/components/schemas/InterviewAnswer"
          minItems: 3
          maxItems: 5
      required: [targetRole, responses]
    MockInterviewEvaluation:
      type: object
      properties:
        source:
          type: string
          enum: [ai, local]
        targetRole:
          type: string
        targetCompany:
          type: string
        overallScore:
          type: integer
          minimum: 0
          maximum: 100
        clarityScore:
          type: integer
          minimum: 0
          maximum: 100
        technicalAccuracy:
          type: integer
          minimum: 0
          maximum: 100
        conceptsCovered:
          type: array
          items:
            type: string
            maxLength: 100
          maxItems: 10
        missedConcepts:
          type: array
          items:
            type: string
            maxLength: 100
          maxItems: 10
        summary:
          type: string
          maxLength: 700
        tips:
          type: array
          items:
            type: string
            maxLength: 320
          minItems: 2
          maxItems: 8
      required: [source, targetRole, targetCompany, overallScore, clarityScore, technicalAccuracy, conceptsCovered, missedConcepts, summary, tips]
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]