# lib/api-zod/src/generated/types/interviewAnswer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewAnswer {
  /** @maxLength 500 */
  question: string;
  /** @maxLength 5000 */
  answer: string;
}