# lib/api-zod/src/generated/types/learningResource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LearningResource {
  /** @maxLength 140 */
  title: string;
  /** @maxLength 100 */
  provider: string;
  /** @maxLength 220 */
  reason: string;
  free: boolean;
}