# lib/api-zod/src/generated/types/mockInterviewEvaluation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MockInterviewEvaluationSource } from './mockInterviewEvaluationSource';

export interface MockInterviewEvaluation {
  source: MockInterviewEvaluationSource;
  targetRole: string;
  targetCompany: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  overallScore: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  clarityScore: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  technicalAccuracy: number;
  /**
     * @maxItems 10
     * @items.maxLength 100
     */
  conceptsCovered: string[];
  /**
     * @maxItems 10
     * @items.maxLength 100
     */
  missedConcepts: string[];
  /** @maxLength 700 */
  summary: string;
  /**
     * @minItems 2
     * @maxItems 8
     * @items.maxLength 320
     */
  tips: string[];
}