# lib/api-zod/src/generated/types/mockInterviewEvaluationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewAnswer } from './interviewAnswer';

export interface MockInterviewEvaluationInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  targetRole: string;
  /** @maxLength 100 */
  targetCompany?: string;
  /**
     * @minItems 3
     * @maxItems 5
     */
  responses: InterviewAnswer[];
}