# lib/api-zod/src/generated/types/mockInterviewQuestions.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MockInterviewQuestionsSource } from './mockInterviewQuestionsSource';

export interface MockInterviewQuestions {
  source: MockInterviewQuestionsSource;
  targetRole: string;
  targetCompany: string;
  /**
     * @minItems 3
     * @maxItems 5
     * @items.maxLength 500
     */
  questions: string[];
}