# lib/api-zod/src/generated/types/mockInterviewQuestionsInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MockInterviewQuestionsInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  targetRole: string;
  /** @maxLength 100 */
  targetCompany?: string;
}