# lib/api-zod/src/generated/types/roadmapPhase.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LearningResource } from './learningResource';
import type { RoadmapPhasePhase } from './roadmapPhasePhase';
import type { WeeklyMilestone } from './weeklyMilestone';

export interface RoadmapPhase {
  phase: RoadmapPhasePhase;
  /** @maxLength 100 */
  title: string;
  /** @maxLength 220 */
  focus: string;
  /**
     * @minItems 4
     * @maxItems 4
     */
  weeklyMilestones: WeeklyMilestone[];
  /** @maxLength 360 */
  portfolioProject: string;
  /**
     * @minItems 2
     * @maxItems 3
     */
  resources: LearningResource[];
  /**
     * @minItems 1
     * @maxItems 3
     * @items.maxLength 260
     */
  resumeTips: string[];
}