# lib/api-zod/src/generated/types/skillGap.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SkillGap {
  /** @maxLength 80 */
  skill: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  current: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  required: number;
  /** @maxLength 320 */
  rationale: string;
  /**
     * @maxItems 6
     * @items.maxLength 80
     */
  missingTools?: string[];
}