# lib/api-zod/src/generated/types/skillLevelInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SkillLevelInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  skill: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  current: number;
}