# lib/api-zod/src/generated/types/skillTwinAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoadmapPhase } from './roadmapPhase';
import type { SkillGap } from './skillGap';
import type { SkillTwinAnalysisSource } from './skillTwinAnalysisSource';

export interface SkillTwinAnalysis {
  targetRole: string;
  source: SkillTwinAnalysisSource;
  /**
     * @minimum 0
     * @maximum 100
     */
  readiness: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  matchScore: number;
  /** @maxLength 800 */
  summary: string;
  /**
     * @minItems 1
     * @maxItems 6
     * @items.maxLength 320
     */
  strengths: string[];
  /**
     * @minItems 1
     * @maxItems 6
     * @items.maxLength 320
     */
  shortfalls: string[];
  /**
     * @minItems 5
     * @maxItems 7
     */
  gaps: SkillGap[];
  /**
     * @minItems 3
     * @maxItems 3
     */
  roadmap: RoadmapPhase[];
}