# lib/api-zod/src/generated/types/skillTwinAnalysisInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SkillLevelInput } from './skillLevelInput';

export interface SkillTwinAnalysisInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  targetRole: string;
  /** @maxLength 16000 */
  resumeText?: string;
  /**
     * @minItems 5
     * @maxItems 7
     */
  skillLevels?: SkillLevelInput[];
}