# lib/api-zod/src/generated/types/weeklyMilestone.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WeeklyMilestone {
  /** @maxLength 30 */
  week: string;
  /** @maxLength 320 */
  goal: string;
  /** @maxLength 220 */
  deliverable: string;
}